secure_share_backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
cryptography==41.0.7
python-jose==3.3.0
passlib==1.7.4
redis==5.0.1
celery==5.3.4
python-dotenv==1.0.0
aioredis==2.0.1